feat: implement error handling for bikes, customers, and services

+++ src/app/modules/customer/customer.service.ts
@@ -1,7 +1,9 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import httpStatus from 'http-status';
 
 import { PrismaClient } from '../../../../prisma/generated/client';
+import { AppError } from '../../errors/AppError';
 
 const prisma = new PrismaClient();
 
@@ -20,6 +22,10 @@ const getSingleCustomerFromDB = async (id: string): Promise<any> => {
     },
   });
 
+  if (!result) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Customer not found');
+  }
+
   return result;
 };
 
@@ -34,6 +40,16 @@ const createCustomerIntoDB = async (data: any): Promise<any> => {
 
 // update customer into db
 const updateCustomerIntoDB = async (id: string, payload: any): Promise<any> => {
+  const isExist = await prisma.customer.findUnique({
+    where: {
+      customerId: id,
+    },
+  });
+
+  if (!isExist) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Customer not found');
+  }
+
   const result = await prisma.customer.update({
     where: {
       customerId: id,
@@ -46,7 +62,15 @@ const updateCustomerIntoDB = async (id: string, payload: any): Promise<any> => {
 
 // delete customer from db
 const deleteCustomerFromDB = async (id: string): Promise<any> => {
-  console.log(id);
+  const isExist = await prisma.customer.findUnique({
+    where: {
+      customerId: id,
+    },
+  });
+
+  if (!isExist) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Customer not found');
+  }
 
   const result = await prisma.customer.delete({
     where: {
